[TensorExpr] Add plumbing for conv2d fusion.
For now the only way to represent conv2d in TE is via an external call,
and since aten library doesn't have an out variant for conv2d, the
external call has to perform an extra copy. Because of that fusing
conv2d now regressed performance and hence is disabled. However, in near
future we should have two alternative ways to enable it:
1) represent conv2d natively in TE (without an external call)
2) add an out variant for conv2d

ghstack-source-id: 317550f5d9a48d2d710ebc9087edd4128ea906de
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54439