[Futures] enhance error handling in then()
Implements the proposal in https://github.com/pytorch/pytorch/issues/53717#issuecomment-800545655. See that issue for more details, but at a high level:

1. markCompleted() immediately sets completed_ = true
2. Subclasses of future (such as cuda future) implement a nontrivial `postMarkCompletedHook` which may throw
3. If above error is caught and we call `setError`, setError itself will error out because completed_ = true.

To fix this, only call setError if the user-defined cb resulted in an error, otherwise, call `markCompleted` and let postMarkCompletedHook() throw and crash the program (per @lw's thoughts this should be a fatal).

Differential Revision: [D27252965](https://our.internmc.facebook.com/intern/diff/D27252965/)

ghstack-source-id: 124601630
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54475