[NCCL] Enhance watchdog to log exceptions
When looping through the nccl communicator cache checking for errors, enhance the watchdog to log exceptions that are set on the communicator.

This will allow for better debugability since the NCCL error will be logged when the watchdog receives errors for the communicators and aborts them appropriately.

Tested by forcing a NCCL error with NCCL_BLOCKING_WAIT=1 and verifying that the exception is indeed logged.

Differential Revision: [D27106699](https://our.internmc.facebook.com/intern/diff/D27106699/)

ghstack-source-id: 124709044
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54557