[NCCL][Blocking Wait] Log set exceptions when checking for exceptions in blocking wait
In blocking wait's polling synchronization loop, we frequently call checkAndSetException() as part of isCompleted() to check the status of nccl operations. It would be useful to log here in case we encounter any exceptions (which are later thrown by `checkAndThrowException`).

Also slightly refactors code previously added to make use of a helper function to get the error message given an `std::exception_ptr`.

Differential Revision: [D27136202](https://our.internmc.facebook.com/intern/diff/D27136202/)

ghstack-source-id: 124709043
Pull Request resolved: https://github.com/pytorch/pytorch/pull/54558